feat: extract InteractiveTimeline event handlers to useTimelineEvents hook (closes #62)
- Create useTimelineEvents hook with extracted event handling logic (~700 lines)
- Extract handleMouseDown, handleMouseMove, handleMouseUp handlers
- Extract handleDoubleClick, handleRightClick handlers
- Extract handlePhaseAction, handlePhaseHandleMouseDown handlers
- Extract tooltip and hover state management
- Maintain full event handling functionality

Results:
- InteractiveTimeline.tsx reduced from 1,750 to 1,223 lines (527 line reduction)
- Exceeds target of ~300 line reduction
- Component now under 1,300 lines (below target of 1,200)
